refactor(admin): remove unused denyBlogAccess wrapper

+++ src/Controller/Admin/PostController.php
@@ -44,7 +44,7 @@ class PostController extends AbstractDomainController
 
     public function indexAction(Request $request): Response
     {
-        $this->denyBlogAccess(BlogPermission::READ, BlogPermission::RES_POST);
+        $this->denyAccessUnlessGranted(BlogPermission::READ, BlogPermission::RES_POST);
 
         $definition = $this->definitionFactory->build();
         $data = $this->dataFactory->build($this->getContextShopId(), $this->getContextLangId(), $request->query->all());
@@ -58,7 +58,7 @@ class PostController extends AbstractDomainController
 
     public function formAction(Request $request, ?int $postId = null): Response
     {
-        $this->denyBlogAccess($postId ? BlogPermission::UPDATE : BlogPermission::CREATE, BlogPermission::RES_POST);
+        $this->denyAccessUnlessGranted($postId ? BlogPermission::UPDATE : BlogPermission::CREATE, BlogPermission::RES_POST);
 
         $form = $this->createForm(PostType::class, $this->formDataProvider->getData($postId));
         $form->handleRequest($request);
@@ -72,7 +72,7 @@ class PostController extends AbstractDomainController
 
     public function createAction(Request $request): JsonResponse
     {
-        $this->denyBlogAccess(BlogPermission::CREATE, BlogPermission::RES_POST);
+        $this->denyAccessUnlessGranted(BlogPermission::CREATE, BlogPermission::RES_POST);
         $this->validateCsrfToken($request, 'everpsblog_post_create');
 
         $command = $this->commandAssembler->assembleCreate($request->request->all());
@@ -83,7 +83,7 @@ class PostController extends AbstractDomainController
 
     public function updateAction(int $postId, Request $request): JsonResponse
     {
-        $this->denyBlogAccess(BlogPermission::UPDATE, BlogPermission::RES_POST);
+        $this->denyAccessUnlessGranted(BlogPermission::UPDATE, BlogPermission::RES_POST);
         $this->validateCsrfToken($request, 'everpsblog_post_update_' . $postId);
 
         $command = $this->commandAssembler->assembleUpdate($postId, $request->request->all());
@@ -95,7 +95,7 @@ class PostController extends AbstractDomainController
 
     public function deleteAction(int $postId, Request $request): JsonResponse
     {
-        $this->denyBlogAccess(BlogPermission::DELETE, BlogPermission::RES_POST);
+        $this->denyAccessUnlessGranted(BlogPermission::DELETE, BlogPermission::RES_POST);
         $this->validateCsrfToken($request, 'everpsblog_post_delete_' . $postId);
 
         $this->commandBus->handle(new DeletePostCommand($postId));
